+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 BUG FIXES:
 
   - core: push configurations are validated with global configs [GH-5130]
-  - guests/funtoo: Fix incorrect path in configure networks [GH-4812]
+  - guests/funtoo: fix incorrect path in configure networks [GH-4812]
+  - plugins/login: allow users to login with a token [GH-5145]
 
 ## 1.7.2 (January 6, 2015)