+++ CHANGELOG.md
@@ -3,11 +3,13 @@
 BUG FIXES:
 
   - core: push configurations are validated with global configs [GH-5130]
+  - core: remove executable permissions on internal file [GH-5220]
   - guests/funtoo: fix incorrect path in configure networks [GH-4812]
   - plugins/login: allow users to login with a token [GH-5145]
   - providers/hyperv: allow users to configure memory, cpu count, and vmname [GH-5183]
   - provisioners/ansible: fix SSH settings to support more than 5 ssh keys [GH-5017]
   - provisioners/ansible: increase ansible connection timeout to 30 seconds [GH-5018]
+  - provisioners/docker: use docker.com instead of docker.io [GH-5216]
 
 ## 1.7.2 (January 6, 2015)